+++ napari/layers/base/__init__.py
@@ -1,1 +1,1 @@
-from .model import Layer
+from .base import Layer
