+++ napari/layers/base/base.py
