+++ napari/layers/image/__init__.py
@@ -1,1 +1,1 @@
-from .model import Image
+from .image import Image
