+++ napari/layers/image/image.py
