+++ napari/layers/labels/labels.py
