+++ napari/layers/points/points.py
