+++ napari/layers/pyramid/pyramid.py
