+++ napari/layers/shapes/shapes.py
