+++ napari/layers/vectors/vectors.py
